extracts custom hooks from App.js
